Some cleanup: breaking into files and comments.
